Move the `Marketplace::Policy::Factories` module to `Spec`
- https://github.com/zinc-collective/convene/issues/1657

It felt really weird to me to have a module calling `let` from within
the `app` directory; since that's very much an `rspec`-thing.

This pulls it out to it's own place; ensures `solargraph` can find the
class definition, comments it, and uses it almost everywhere
`let(:marketplace)` is used.

Again, not sure it's a great / good idea...
