+++ app/furniture/marketplace/policy.rb
@@ -23,17 +23,5 @@ class Marketplace
 
       object.marketplace if object.respond_to?(:marketplace)
     end
-
-    module SpecFactories
-      def self.included(spec)
-        spec.let(:marketplace) { create(:marketplace) }
-        spec.let(:membership) { create(:membership, space: marketplace.room.space) }
-        spec.let(:member) { membership.member.becomes(Person) }
-        spec.let(:neighbor) { create(:marketplace_person) }
-        spec.let(:operator) { create(:marketplace_person, operator: true) }
-        spec.let(:guest) { Guest.new(session: session) }
-        spec.let(:session) { {} }
-      end
-    end
   end
 end
