+++ spec/furniture/marketplace/marketplace_policy_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.describe Marketplace::MarketplacePolicy, type: :policy do
   subject { described_class }
 
-  include Marketplace::Policy::SpecFactories
+  include Spec::Marketplace::CommonLets
 
   permissions :index?, :show? do
     it { is_expected.to permit(member, marketplace) }
