+++ spec/furniture/marketplace/onboarding_component_spec.rb
@@ -3,8 +3,7 @@ require "rails_helper"
 RSpec.describe Marketplace::OnboardingComponent, type: :component do
   subject(:output) { render_inline(component) }
 
-  let(:marketplace) { create(:marketplace) }
-  let(:operator) { build(:person, operator: true) }
+  include Spec::Marketplace::CommonLets
 
   let(:component) { described_class.new(marketplace: marketplace, current_person: operator) }
