+++ spec/furniture/marketplace/order/received_mailer_spec.rb
@@ -1,7 +1,8 @@
 require "rails_helper"
 
 RSpec.describe Marketplace::Order::ReceivedMailer, type: :mailer do
-  let(:marketplace) { create(:marketplace) }
+  include Spec::Marketplace::CommonLets
+
   let(:order) { build(:marketplace_order, placed_at: 1.hour.ago, marketplace: marketplace) }
 
   describe "#notification" do
