+++ spec/furniture/marketplace/order_policy_spec.rb
@@ -3,7 +3,7 @@ require "rails_helper"
 RSpec.describe Marketplace::OrderPolicy, type: :policy do
   subject { described_class }
 
-  include Marketplace::Policy::SpecFactories
+  include Spec::Marketplace::CommonLets
 
   let(:order) { create(:marketplace_order, shopper: shopper, marketplace: marketplace) }
   let(:shopper) { build(:marketplace_shopper) }
